Bump and reorganize FoodCritic
In order to prepare a more significant switch to ChefDK, the FoodCritic
exceptions are now defined in .foodcritic, and no longer in the
Rakefile.

Enable back FC041 (download are done via the ark cookbook only)

Tolerate FC053 as the 'recommends' keyword is still worth to be present,
even though Chef tools are not using this meta data.