[GH-34] Always download .zip packages with new URL
After four random checks, it seems that the new HTTPS-based
repository provides all the Linux package history. The archives are in
.zip format, and don't require to strip an `android-sdk-linux` base
directory.

Since the new .zip archives only contain the 'tools' directory, it is
no longer necessary to fix the permissions and ownership for
the 'add-ons' or 'platforms' that were present in the .tgz file.